+++ tests/sources/test_document_oereblex.py
@@ -138,49 +138,49 @@ def test_get_document_records(i, document):
 
 def test_read():
     with requests_mock.mock() as m:
-        with open('./tests/resources/geolink_v1.1.1.xml', 'rb') as f:
+        with open('./tests/resources/geolink_v1.2.0.xml', 'rb') as f:
             m.get('http://oereblex.example.com/api/geolinks/100.xml', content=f.read())
         source = OEREBlexSource(host='http://oereblex.example.com', language='de', canton='BL')
         source.read(MockParameter(), 100)
-        assert len(source.records) == 2
+        assert len(source.records) == 5
         document = source.records[0]
         assert isinstance(document, DocumentRecord)
         assert isinstance(document.responsible_office, OfficeRecord)
-        assert document.responsible_office.name == {'de': 'Landeskanzlei'}
+        assert document.responsible_office.name == {'de': 'Bauverwaltung Gemeinde'}
         assert document.canton == 'BL'
         assert document.text_at_web == {
-            'de': 'http://oereblex.example.com/api/attachments/313'
+            'de': 'http://oereblex.example.com/api/attachments/4735'
         }
-        assert len(document.references) == 5
+        assert len(document.references) == 4
 
 
 def test_read_related_decree_as_main():
     with requests_mock.mock() as m:
-        with open('./tests/resources/geolink_v1.1.1.xml', 'rb') as f:
+        with open('./tests/resources/geolink_v1.2.0.xml', 'rb') as f:
             m.get('http://oereblex.example.com/api/geolinks/100.xml', content=f.read())
         source = OEREBlexSource(host='http://oereblex.example.com', language='de', canton='BL',
                                 related_decree_as_main=True)
         source.read(MockParameter(), 100)
-        assert len(source.records) == 3
+        assert len(source.records) == 5
         document = source.records[0]
         assert isinstance(document, DocumentRecord)
         assert isinstance(document.responsible_office, OfficeRecord)
-        assert document.responsible_office.name == {'de': 'Landeskanzlei'}
+        assert document.responsible_office.name == {'de': 'Bauverwaltung Gemeinde'}
         assert document.canton == 'BL'
         assert document.text_at_web == {
-            'de': 'http://oereblex.example.com/api/attachments/313'
+            'de': 'http://oereblex.example.com/api/attachments/4735'
         }
         assert len(document.references) == 4
 
 
 def test_read_with_version_in_url():
     with requests_mock.mock() as m:
-        with open('./tests/resources/geolink_v1.1.1.xml', 'rb') as f:
-            m.get('http://oereblex.example.com/api/1.1.1/geolinks/100.xml', content=f.read())
+        with open('./tests/resources/geolink_v1.2.0.xml', 'rb') as f:
+            m.get('http://oereblex.example.com/api/1.2.0/geolinks/100.xml', content=f.read())
         source = OEREBlexSource(host='http://oereblex.example.com', language='de', canton='BL',
                                 pass_version=True)
         source.read(MockParameter(), 100)
-        assert len(source.records) == 2
+        assert len(source.records) == 5
 
 
 def test_read_with_specified_version():
@@ -195,17 +195,17 @@ def test_read_with_specified_version():
 
 def test_read_with_specified_language():
     with requests_mock.mock() as m:
-        with open('./tests/resources/geolink_v1.1.1.xml', 'rb') as f:
+        with open('./tests/resources/geolink_v1.2.0.xml', 'rb') as f:
             m.get('http://oereblex.example.com/api/geolinks/100.xml?locale=fr', content=f.read())
         source = OEREBlexSource(host='http://oereblex.example.com', language='de', canton='BL')
         params = MockParameter()
         params.set_language('fr')
         source.read(params, 100)
-        assert len(source.records) == 2
+        assert len(source.records) == 5
         document = source.records[0]
-        assert document.responsible_office.name == {'fr': 'Landeskanzlei'}
+        assert document.responsible_office.name == {'fr': 'Bauverwaltung Gemeinde'}
         assert document.text_at_web == {
-            'fr': 'http://oereblex.example.com/api/attachments/313'
+            'fr': 'http://oereblex.example.com/api/attachments/4735'
         }