+++ doc/source/changes.rst
@@ -12,7 +12,14 @@ to adapt in your project configuration, database etc. when upgrading to a new ve
 Version 1.7.2 (CURRENT DEV)
 ---------------------------
 This is the current development version, not yet released.
-Changes thus far are relevant only for users of MapFish Print.
+
+Oereblex related changes
+^^^^^^^^^^^^^^^^^^^^^^^^
+Pyramid Oereb now supports and uses by default geoLink schema version 1.2.0:
+
+ * New doctype 'notice' (will be classified as 'HintRecord'). 'Notice' can have no authority nor enactment_date
+ * New document attribute 'language' and file attribute 'description' are currently not used by
+   pyramid_oereb but are now available to custom code, for example for document title generation.
 
 MapFish Print related changes
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^